[CHORE] Alterações no endpoint de prestação de serviço para registrar o id do usuário logado no banco

+++ back-facilita_ai/src/main/java/com/UFMSPetSistemas/getpet/controller/fotos/ArquivoController.java
@@ -4,6 +4,7 @@ import com.UFMSPetSistemas.getpet.model.entities.Usuario;
 import com.UFMSPetSistemas.getpet.model.repository.UsuarioRepository;
 import com.UFMSPetSistemas.getpet.services.MinioService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -30,19 +31,31 @@ public class ArquivoController {
             @AuthenticationPrincipal Usuario usuario,
             @RequestPart("file") MultipartFile file
     ) {
+        try {
 
-        if (usuario.getFotoPerfil() != null) {
-            minioService.delete(usuario.getFotoPerfil());
+            try {
+                if (usuario.getFotoPerfil() != null) {
+                    String oldObjectName = minioService.extractObjectName(usuario.getFotoPerfil());
+                    if (oldObjectName != null) {
+                        minioService.delete(usuario.getFotoPerfil());
+                    }
+                }
+            } catch (Exception e) {
+                System.err.println("Erro: " + e.getMessage());
+            }
+
+            String objectName = "usuarios/perfil/" + usuario.getId() + UUID.randomUUID();
+            minioService.upload(file, objectName);
+
+            String url = objectName;
+            usuario.setFotoPerfil(url);
+            usuarioRepo.save(usuario);
+
+            return ResponseEntity.ok(new UploadResponseDTO(url));
+        } catch (Exception e) {
+            e.printStackTrace();
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
         }
-
-        String objectName = "usuarios/perfil/" + usuario.getId() + UUID.randomUUID();
-        minioService.upload(file, objectName);
-
-        String url = minioService.generatePresignedUrl(objectName);
-        usuario.setFotoPerfil(url);
-        usuarioRepo.save(usuario);
-
-        return ResponseEntity.ok(new UploadResponseDTO(url));
     }
 
     @GetMapping("/view/profile")
@@ -51,8 +64,13 @@ public class ArquivoController {
             if (usuario.getFotoPerfil() == null) {
                 return ResponseEntity.notFound().build();
             }
-
             String objectName = minioService.extractObjectName(usuario.getFotoPerfil());
+
+            if (objectName == null) {
+                System.err.println("URL inválida ou corrompida, não foi possível extrair nome do arquivo.");
+                return ResponseEntity.notFound().build();
+            }
+
             byte[] bytes = minioService.getObject(objectName);
             String contentType = minioService.getContentType(objectName);
 
@@ -61,6 +79,7 @@ public class ArquivoController {
                     .body(bytes);
 
         } catch (Exception e) {
+            e.printStackTrace();
             return ResponseEntity.status(500).build();
         }
     }
